requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
passlib[bcrypt]
python-jose
python-multipart
youtube-search
bcrypt==3.2.0